541	Change the distance type in java to int

+++ npbenchmark-main/ACS_VRPTW/src/vrptw/AntColonySystem.java
@@ -25,7 +25,7 @@ public class AntColonySystem
     // Dimension: num__nodes * num__nodes == (客户数量+1) * (客户数量+1)
     // Meaning: distance[i][j], 点i和点j的距离;
     // Comment: 这也是为什么长度选num__nodes而不是num__customers的原因, 因为仓库和各点之间的距离也需要计算;
-    double[][] distance;
+    int[][] distance;
 
     // Dimension: num__agents == num__customers == num__nodes - 1;
     // untreated[i]: 记录agent i尚未服务过的客户
@@ -117,8 +117,8 @@ public class AntColonySystem
                 {
                     pheromone[i][j] = init_pheromone;
                     pheromone[j][i] = init_pheromone;
-                    heuristic[i][j] = 1 / distance[i][j];
-                    heuristic[j][i] = 1 / distance[i][j];
+                    heuristic[i][j] = 1 / (double) distance[i][j];
+                    heuristic[j][i] = 1 / (double) distance[i][j];
                 }
             }
         }
